+++ lib/hyperclient/entry_point.rb
@@ -29,6 +29,7 @@ module Hyperclient
   #
   class EntryPoint < Link
     extend Forwardable
+
     # Public: Delegates common methods to be used with the Faraday connection.
     def_delegators :connection, :basic_auth, :digest_auth, :token_auth, :headers, :headers=, :params, :params=